add nssdb::add_cert define for importing certs into a db
In addition, this changeset is removing this functionality and the
cacert, canickname, and catrust params from the nssdb::create define.
It would be easy to add support back to that define as a wrapper around
the new nssdb::add_cert type or to add a new 'convenience' type to
accomplish the same thing.